+++ src/process_binance_com.rs
@@ -18,7 +18,7 @@ use crate::{
     configuration::Configuration,
     de_string_to_utc_time_ms::{de_string_to_utc_time_ms_condaddtzutc, se_time_ms_to_utc_string},
     token_tax_comment_vers::{TT_CMT_VER1, TT_CMT_VER2},
-    token_tax_processing::{TokenTaxRec, TypeTxs},
+    process_token_tax::{TokenTaxRec, TypeTxs},
 };
 use clap::ArgMatches;
 use rust_decimal::prelude::*;
@@ -1030,7 +1030,7 @@ fn write_tr_vec(
 mod test {
 
     use super::*;
-    use crate::token_tax_processing::{TokenTaxRec, TypeTxs};
+    use crate::process_token_tax::{TokenTaxRec, TypeTxs};
     use rust_decimal_macros::dec;
 
     #[test]
